Conflicts:
	xorg-server/glx/dispatch.h
	xorg-server/glx/glapi.c
	xorg-server/glx/indirect_dispatch.h
	xorg-server/glx/indirect_reqsize.c
	xorg-server/glx/indirect_reqsize.h
	xorg-server/glx/indirect_size.h
	xorg-server/glx/indirect_size_get.c
	xorg-server/glx/indirect_size_get.h
	xorg-server/glx/indirect_table.c
	xorg-server/mi/mieq.c

+++ xorg-server/hw/kdrive/linux/ps2.c
@@ -159,7 +159,7 @@ Ps2Enable (KdPointerInfo *pi)
         return BadAlloc;
     }
 
-    pi->driverPrivate = (void *)fd;
+    pi->driverPrivate = (void *)(intptr_t)fd;
 
     return Success;
 }
@@ -168,7 +168,7 @@ Ps2Enable (KdPointerInfo *pi)
 static void
 Ps2Disable (KdPointerInfo *pi)
 {
-    KdUnregisterFd (pi, (int)pi->driverPrivate, TRUE);
+    KdUnregisterFd (pi, (int)(intptr_t)pi->driverPrivate, TRUE);
 }
 
 static void
